+++ cmd/gravinet/main.go
@@ -48,7 +48,7 @@ import (
 
 // Build metadata, overridable via -ldflags.
 var (
-	version = "613"
+	version = "614"
 	commit  = "none"
 )